requirements.txt
mysql-connector-python
python-dotenv
langchain
langchain-community
langchain-experimental
langchain-openai
